This is a report about a time series.
First, I used an ARIMA model. This model looks at the past data to find patterns. I had to make the data steady first. The model worked, but I wanted to see if another model was better.

Then, I used a double exponential model. This is another way to predict things, especially for data with a trend. It made its own predictions.

I compared the two models. I used two scores, AIC and SBC. The double exponential model had a lower score on both. This means it was the better model. It was the best choice for forecasting the future of this data.